requirements.txt
streamlit
matplotlib
seaborn
pandas
numpy
nltk
scikit-learn
transformers
sentence-transformers
faiss-cpu
wordcloud
urlextract
emoji
torch
